Move helper
We want to use this helper for the new concept of 'project groups' so
we'll move it somewhere more generic.
